CRUD Crancial and Rack and Image and CSS

+++ Controllers/RackSamplesController.cs
@@ -6,36 +6,53 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using IWantMyMummy.Models;
+using IWantMyMummy.Models.ViewModels;
+using IWantMyMummy.Data;
+using IWantMyMummy.Areas.Identity.Data;
+using Microsoft.AspNetCore.Identity;
+
 
 namespace IWantMyMummy.Controllers
 {
     public class RackSamplesController : Controller
     {
         private readonly MummyContext _context;
+        private IWantMyMummyContext mummyContext;
+        private UserManager<IWantMyMummyUser> userManager;
 
-        public RackSamplesController(MummyContext context)
+        public RackSamplesController(MummyContext context, IWantMyMummyContext con, UserManager<IWantMyMummyUser> tempUser)
         {
             _context = context;
+            mummyContext = con;
+            userManager = tempUser;
+
         }
 
         // GET: RackSamples
         public async Task<IActionResult> Index()
         {
-            var mummyContext = _context.RackSample.Include(r => r.Burial);
-            return View(await mummyContext.ToListAsync());
+            var role = (mummyContext.UserRoles
+           .Where(r => r.UserId == userManager.GetUserId(User))
+           .FirstOrDefault());
+
+            if (!(role is null))
+            {
+                ViewBag.Role = Int32.Parse(role.RoleId);
+            }
+            var mumCon = _context.RackSample.Include(r => r.Burial);
+            return View(await mumCon.ToListAsync());
         }
 
         // GET: RackSamples/Details/5
-        public async Task<IActionResult> Details(string id)
+        public async Task<IActionResult> Details(string RackNumber, string RackShelf)
         {
-            if (id == null)
+            if (RackNumber == null || RackShelf==null)
             {
                 return NotFound();
             }
 
-            var rackSample = await _context.RackSample
-                .Include(r => r.Burial)
-                .FirstOrDefaultAsync(m => m.RackShelf == id);
+            var rackSample = _context.RackSample.Where(x => x.RackNumber == RackNumber && x.RackShelf == RackShelf).First();
+                
             if (rackSample == null)
             {
                 return NotFound();
@@ -45,10 +62,28 @@ namespace IWantMyMummy.Controllers
         }
 
         // GET: RackSamples/Create
-        public IActionResult Create()
-        {
+        public IActionResult Create(int BurialId, string Burial, string Subplot, int Num)
+        {   
+            if (BurialId <= 0)
+            {
+                return NotFound();
+            }
+            ViewBag.Id = BurialId;
+            ViewBag.Burial = Burial;
+            ViewBag.Subplot = Subplot;
+            ViewBag.Num = Num;
+            RackViewModel viewModel = new RackViewModel
+            {
+                BurialId = BurialId,
+                RackSample = new RackSample
+                {
+                    BurialId = BurialId,
+                },
+                BurialList = _context.Burial.Where(x => x.BurialId == BurialId).ToList(),
+            };
+            
             ViewData["BurialId"] = new SelectList(_context.Burial, "BurialId", "BurialWrapping");
-            return View();
+            return View(viewModel);
         }
 
         // POST: RackSamples/Create
@@ -69,14 +104,15 @@ namespace IWantMyMummy.Controllers
         }
 
         // GET: RackSamples/Edit/5
-        public async Task<IActionResult> Edit(string id)
+        public async Task<IActionResult> Edit(string RackNumber, string RackShelf)
         {
-            if (id == null)
+            if (RackNumber == null || RackShelf == null)
             {
                 return NotFound();
             }
 
-            var rackSample = await _context.RackSample.FindAsync(id);
+            var rackSample = _context.RackSample.Where(x => x.RackNumber == RackNumber && x.RackShelf == RackShelf).First();
+
             if (rackSample == null)
             {
                 return NotFound();
@@ -90,9 +126,9 @@ namespace IWantMyMummy.Controllers
         // more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(string id, [Bind("RackShelf,RackNumber,IsBag,BurialId,TubeNumber,RankDescription,MlSize,Foci,LocationNotes,Questions,Conventional14cAgeBp,_14cCalendarDay,Calibrated95CalendarDateMax,Calibrated95CalendarDateMin,Calibrated95CalendarDateSpan,Calibrated95CalendarDateAvg,Category,Notes")] RackSample rackSample)
+        public async Task<IActionResult> Edit(string RackNumber, string RackShelf, [Bind("RackShelf,RackNumber,IsBag,BurialId,TubeNumber,RankDescription,MlSize,Foci,LocationNotes,Questions,Conventional14cAgeBp,_14cCalendarDay,Calibrated95CalendarDateMax,Calibrated95CalendarDateMin,Calibrated95CalendarDateSpan,Calibrated95CalendarDateAvg,Category,Notes")] RackSample rackSample)
         {
-            if (id != rackSample.RackShelf)
+            if (RackShelf != rackSample.RackShelf || RackNumber != rackSample.RackNumber)
             {
                 return NotFound();
             }
@@ -122,16 +158,15 @@ namespace IWantMyMummy.Controllers
         }
 
         // GET: RackSamples/Delete/5
-        public async Task<IActionResult> Delete(string id)
+        public async Task<IActionResult> Delete(string RackNumber, string RackShelf)
         {
-            if (id == null)
+            if (RackNumber == null || RackShelf == null)
             {
                 return NotFound();
             }
 
-            var rackSample = await _context.RackSample
-                .Include(r => r.Burial)
-                .FirstOrDefaultAsync(m => m.RackShelf == id);
+            var rackSample = _context.RackSample.Where(x => x.RackNumber == RackNumber && x.RackShelf == RackShelf).First();
+
             if (rackSample == null)
             {
                 return NotFound();
@@ -143,9 +178,20 @@ namespace IWantMyMummy.Controllers
         // POST: RackSamples/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(string id)
+        public async Task<IActionResult> DeleteConfirmed(string RackNumber, string RackShelf)
         {
-            var rackSample = await _context.RackSample.FindAsync(id);
+            if (RackNumber == null || RackShelf == null)
+            {
+                return NotFound();
+            }
+
+            var rackSample = _context.RackSample.Where(x => x.RackNumber == RackNumber && x.RackShelf == RackShelf).First();
+
+            if (rackSample == null)
+            {
+                return NotFound();
+            }
+
             _context.RackSample.Remove(rackSample);
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
